+++ stripe_terminal/lib/mek_stripe_terminal.dart
@@ -3,6 +3,7 @@
 library mek_stripe_terminal;
 
 export 'src/cancellable_future.dart';
+export 'src/models/card.dart';
 export 'src/models/cart.dart';
 export 'src/models/discover_config.dart';
 export 'src/models/location.dart';
@@ -11,6 +12,8 @@ export 'src/models/payment_intent.dart';
 export 'src/models/payment_method.dart';
 export 'src/models/reader.dart';
 export 'src/models/reader_software_update.dart';
+export 'src/models/refund.dart';
+export 'src/models/setup_intent.dart';
 export 'src/reader_delegates.dart';
 export 'src/stripe_terminal.dart';
 export 'src/terminal_exception.dart';